Created settings menu, fixed monster movement
Added the settings menu functionality to the main and pause menu and created an option for mouse sensitivity that saves across play sessions. I did this using PlayerPrefs in a script called GamePreferenceManager

I also made tweaks in the MonsterMovement script to add an offset that the monster will stop at before the player(this should be set very low like 0.05) and turned off collision between the monster and the player so that it can't shove the player through the walls anymore.

I also organized the script folder between menu, player and monster scripts for better accessibility
