+++ Don't Move/Assets/Animations/PauseMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7270b5b0a17cd34991ad5b57d36a126
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
