+++ Don't Move/Assets/Animations/PauseMenu/PauseMenu.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6889a7a1d49dc4543b1ba0943cbc28aa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
