--- Don't Move/Assets/MonsterSoundControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4d68ac85837e794cae81980d49dbb28
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
