--- Don't Move/Assets/PauseCanvas.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6db1f42b6ac0da1448edab0cd5f2f238
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
