--- Don't Move/Assets/PlayAnimation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6bda53e9aebf5b248b1f9ae91c53b026
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
