+++ Don't Move/Assets/Prefabs/SettingsCanvas.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d660a70af86ff8d4ca3c15274c4bf2a6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
