--- Don't Move/Assets/Scripts/FreezeMonsterAnimations.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3aae0b339ba8cb4a8b68f07edbe6e1e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
