+++ Don't Move/Assets/Scripts/Menu Scripts/GamePreferenceManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0470f3a1b9f960488dd3f7f88bd52c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
