+++ Don't Move/Assets/Scripts/Menu Scripts/GameSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7df0e0ef04c6be843b0e246df480a7c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
