+++ Don't Move/Assets/Scripts/Monster Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7b2704f3f2222a40822886e9a0ccf8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
