+++ Don't Move/Assets/Scripts/Monster Scripts/MonsterMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e359bb3904bc42ad8ba4cbfd05c8631
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
