+++ Don't Move/Assets/Scripts/Player Scripts/GrabScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e16990a48f6a43299f0d85836592b37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
