+++ Don't Move/Assets/Scripts/Player Scripts/PlayerOptions.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerOptions : MonoBehaviour
+{
+    
+    
+}
