--- Don't Move/Assets/Scripts/PlayerOptions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 880b3ff6de3004c4dbc0899ce7d212bf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
